ref: add attachments to responses before return
